apps/ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
pdfplumber
python-docx
httpx